+++ CHANGELOG.md
@@ -19,6 +19,8 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Fixed
 
+- Console not applying webhook field masks when creating a webhook from a template that has field masks set.
+
 ### Security
 
 ## [3.25.2] - unreleased